added infinite scroll and middleware
now authorization is required for all actions except get, and the feed is loaded as it scrolls

+++ backend/controllers/comment.controller.js
@@ -29,28 +29,39 @@ class commentController {
  };
 
  createComment = ((req, res) => {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
     Comment.create(req.body)
         .then(result => res.status(200).json({ result }))
         .catch((error) => res.status(500).json({msg:  error }))
 })
 
  updateComment = ((req, res) => {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
     Comment.findOneAndUpdate({ _id: req.params.commentID }, req.body, { new: true, runValidators: true })
         .then(result => res.status(200).json({ result }))
         .catch((error) => res.status(404).json({msg: 'Comment not found' }))
 })
 
  deleteComment = ((req, res) => {
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
     Comment.findOneAndDelete({ _id: req.params.commentID })
         .then(result => res.status(200).json({ result }))
         .catch((error) => res.status(404).json({msg: 'Comment not found' }))
 })
 
 async deleteCommentsFromPost (req, res){
+  if (!req.user) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
   try {
     const postId =  req.params.postId;
     const result = await Comment.deleteMany({ postId });
-    console.log(postId);
     res.status(200).json({ result });
   } catch (error) {
     res.status(404).json({ msg: "Comment not found" });
